Conflicts:
doomsday/engine/api/doomsday.def
doomsday/engine/api/doomsday.h
doomsday/engine/api/sys_audiod.h
doomsday/engine/data/cphelp.txt
doomsday/engine/portable/include/p_bmap.h
doomsday/engine/portable/include/p_maptypes.h
doomsday/engine/portable/include/p_objlink.h
doomsday/engine/portable/include/s_mus.h
doomsday/engine/portable/src/dam_blockmap.c
doomsday/engine/portable/src/dam_file.c
doomsday/engine/portable/src/dam_main.c
doomsday/engine/portable/src/dd_wad.c
doomsday/engine/portable/src/edit_map.c
doomsday/engine/portable/src/m_args.c
doomsday/engine/portable/src/net_main.c
doomsday/engine/portable/src/p_bmap.c
doomsday/engine/portable/src/p_dmu.c
doomsday/engine/portable/src/p_intercept.c
doomsday/engine/portable/src/p_objlink.c
doomsday/engine/portable/src/r_data.c
doomsday/engine/portable/src/r_lumobjs.c
doomsday/engine/portable/src/r_main.c
doomsday/engine/portable/src/r_shadow.c
doomsday/engine/portable/src/rend_main.c
doomsday/engine/portable/src/rend_shadow.c
doomsday/engine/portable/src/s_main.c
doomsday/engine/portable/src/s_mus.c
doomsday/plugins/common/src/p_mapspec.c
doomsday/plugins/common/src/rend_automap.c
doomsday/plugins/jdoom/src/p_enemy.c
doomsday/plugins/jhexen/src/p_inter.c

+++ doomsday/engine/api/sys_audiod_mus.h
@@ -50,7 +50,7 @@ typedef struct audiointerface_music_generic_s {
 // Driver interface for playing music.
 typedef struct audiointerface_music_s {
     audiointerface_music_generic_t gen;
-    void*           (*SongBuffer) (size_t length);
+    void*           (*SongBuffer) (unsigned int length);
     int             (*Play) (int looped);
     int             (*PlayFile) (const char *filename, int looped);
 } audiointerface_music_t;
